extract store,value
